changes for docs, move selector features to features module
