+++ fastplotlib/graphics/_features/__init__.py
@@ -3,6 +3,7 @@ from ._data import PointsDataFeature, ImageDataFeature, HeatmapDataFeature
 from ._present import PresentFeature
 from ._thickness import ThicknessFeature
 from ._base import GraphicFeature, GraphicFeatureIndexable, FeatureEvent, to_gpu_supported_dtype
+from ._selection_features import LinearSelectionFeature, LinearRegionSelectionFeature
 
 __all__ = [
     "ColorFeature",
@@ -17,5 +18,7 @@ __all__ = [
     "GraphicFeature",
     "GraphicFeatureIndexable",
     "FeatureEvent",
-    "to_gpu_supported_dtype"
+    "to_gpu_supported_dtype",
+    "LinearSelectionFeature",
+    "LinearRegionSelectionFeature",
 ]
