+++ fastplotlib/graphics/_features/_thickness.py
@@ -7,14 +7,14 @@ class ThicknessFeature(GraphicFeature):
 
     **event pick info:**
 
-     ===================== ======================== =========================================================================
+     ==================== ======================== =========================================================================
       key                  type                     description
      ==================== ======================== =========================================================================
       "index"              ``None``                 not used
       "new_data"           ``float``                new thickness value
       "collection-index"   int                      the index of the graphic within the collection that triggered the event
       "world_object"       pygfx.WorldObject        world object
-     ==================== ======================== ========================================================================
+     ==================== ======================== =========================================================================
     """
 
     def __init__(self, parent, thickness: float):
